Products2

+++ app/Http/Controllers/Api/Admin/ProductController.php
@@ -10,24 +10,44 @@ class ProductController extends Controller
 {
     public function index()
     {
+        $products = Product::with(['startup', 'subCategory.category', 'images'])->paginate();
 
-        $products = Product::with(['startup', 'images', 'sizes'])->get();
-
+        // Transform the collection
+        $products->getCollection()->transform(function ($product) {
+            $product->sub_category = $product->subCategory;
+            $product->sub_category->category = $product->sub_category->category;
+            unset($product->subCategory);
+            return $product;
+        });
         return response()->success($products);
+
     }
 
     public function show($id)
     {
-        $product = Product::with(['startup', 'images', 'sizes'])->findOrFail($id);
+        $product = Product::with(['startup', 'subCategory.category', 'images'])->find($id);
+
+        if (!$product) {
+            return response()->errors('Product not found');
+        }
+
+        $product->sub_category = $product->subCategory;
+        $product->sub_category->category = $product->sub_category->category;
+        unset($product->subCategory);
 
         return response()->success($product);
     }
 
     public function destroy($id)
     {
-        $product = Product::findOrFail($id);
+        $product = Product::find($id);
+
+        if (!$product) {
+            return response()->errors('Product not found');
+        }
+
         $product->delete();
 
-        return response()->success('product deleted successfully');
+        return response()->success('Product deleted successfully');
     }
 }
